chore: Move color picker input into own file
- Basically the original implementation idea by Copilot.
